+++ src/Page/LoginPage.tsx
@@ -1,63 +1,16 @@
-import { useEffect, useState, useCallback } from "react";
-import { useForm } from "react-hook-form";
 import { LogInIcon } from "lucide-react";
-import { zodResolver } from "@hookform/resolvers/zod";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent } from "@/components/ui/card";
 import { Alert, AlertDescription } from "@/components/ui/alert";
 import { Separator } from "@/components/ui/separator";
-import { login } from "../api/auth/index";
-import { loginFormSchema } from "../model";
-import type { TLoginForm } from "../model";
 import AuthHeader from "@/components/auth/AuthHeader";
 import AuthFooter from "@/components/auth/AuthFooter";
 import LoginForm from "@/components/auth/LoginForm";
 
+import { useLogin } from "../hooks/useLogin";
+
 const LoginPage = () => {
-  const [error, setError] = useState("");
-  const [loading, setLoading] = useState(false);
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors },
-  } = useForm<TLoginForm>({
-    resolver: zodResolver(loginFormSchema), // Zod 스키마 적용
-  });
-
-  const onSubmit = async (formData: TLoginForm) => {
-    try {
-      setLoading(true);
-      const response = await login(formData);
-      if (!response.success) {
-        setError("이메일 또는 비밀번호가 올바르지 않습니다");
-      }
-    } catch (error) {
-      setError("로그인 중 오류가 발생했습니다");
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleGuestLogin = useCallback(async (email: string, password: string) => {
-    try {
-      setLoading(true);
-      const response = await login({ email, password });
-      if (!response.success) {
-        setError("게스트 로그인 실패");
-      }
-    } catch (error) {
-      setError("게스트 로그인 중 오류가 발생했습니다");
-    } finally {
-      setLoading(false);
-    }
-  }, []);
-
-  useEffect(() => {
-    if (!error) return;
-    const timer = setTimeout(() => setError(""), 3000);
-    return () => clearTimeout(timer);
-  }, [error]);
+  const { error, loading, register, errors, handleSubmit, handleGuestLogin } = useLogin();
 
   return (
     <div className="flex min-h-screen items-center justify-center p-4 bg-background">
@@ -65,7 +18,7 @@ const LoginPage = () => {
         <CardContent className="p-6 space-y-6">
           <AuthHeader icon={LogInIcon} title="로그인" />
 
-          <form onSubmit={handleSubmit(onSubmit)} className="space-y-6">
+          <form onSubmit={handleSubmit} className="space-y-6">
             <LoginForm register={register} errors={errors} />
 
             {error && (
